requirements.txt
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
tqdm==4.67.1
pillow==11.1.0
matplotlib==3.10.0
numpy==1.24.4
pandas==2.2.3
opencv-python==4.11.0.86
albumentations==2.0.5
scikit-image==0.25.0
seaborn==0.13.2
fastapi==0.103.0
uvicorn==0.23.2
pyyaml
requests
python-multipart
python-dotenv
psycopg2-binary>=2.9.0